+++ packages/on_audio_query_platform_interface/lib/src/log/log_config.dart
@@ -1,0 +1,16 @@
+import '../types/log_type.dart';
+
+/// Used to configure the logging behavior.
+class LogConfig {
+  /// Used to configure the logging behavior.
+  LogConfig({
+    this.logType = LogType.WARN,
+    this.showDetailedLog = false,
+  });
+
+  /// Define the logging level.
+  final LogType logType;
+
+  /// Define if detailed log will be shown (Disable by default to avoid spam).
+  final bool showDetailedLog;
+}