+++ packages/on_audio_query_platform_interface/lib/src/on_audio_query_core.dart
@@ -1,0 +1,30 @@
+library on_audio_query_core;
+
+//Extensions
+export 'extensions/models_extensions.dart';
+
+//Models
+export 'models/song_model.dart';
+export 'models/album_model.dart';
+export 'models/artist_model.dart';
+export 'models/playlist_model.dart';
+export 'models/genre_model.dart';
+//
+export 'models/device_model.dart';
+
+//Sort Types
+export 'types/sort_types/song_sort_type.dart';
+export 'types/sort_types/album_sort_type.dart';
+export 'types/sort_types/artist_sort_type.dart';
+export 'types/sort_types/playlist_sort_type.dart';
+export 'types/sort_types/genre_sort_type.dart';
+//
+export 'types/order_type.dart';
+export 'types/log_type.dart';
+export 'types/artwork_type.dart';
+export 'types/audios_from_type.dart';
+export 'types/with_filters_type.dart';
+export 'types/uri_type.dart';
+
+//Log
+export 'log/log_config.dart';