+++ packages/on_audio_query_platform_interface/lib/src/types/audios_from_type.dart
@@ -1,0 +1,25 @@
+// ignore_for_file: constant_identifier_names
+
+/// Defines where audios will be acquired.
+enum AudiosFromType {
+  /// Audios from specific Album name.
+  ALBUM,
+
+  /// Audios from specific Album id.
+  ALBUM_ID,
+
+  /// Audios from specific Artist name.
+  ARTIST,
+
+  /// Audios from specific Artist id.
+  ARTIST_ID,
+
+  /// Audios from specific Genre name.
+  GENRE,
+
+  /// Audios from specific Genre id.
+  GENRE_ID,
+
+  /// Audios from specific Playlist.
+  PLAYLIST,
+}