+++ packages/on_audio_query_platform_interface/lib/src/types/sort_types/genre_sort_type.dart
@@ -1,0 +1,7 @@
+// ignore_for_file: constant_identifier_names
+
+/// Defines the sort type used for [queryPlaylists].
+enum GenreSortType {
+  ///[NAME] will return song list based in genre [name].
+  GENRE,
+}