+++ packages/on_audio_query_platform_interface/lib/src/types/sort_types/playlist_sort_type.dart
@@ -1,0 +1,10 @@
+// ignore_for_file: constant_identifier_names
+
+/// Defines the sort type used for [queryPlaylists].
+enum PlaylistSortType {
+  /// [PLAYLIST] will return playlist based in [playlist].
+  PLAYLIST,
+
+  /// [DATE_ADDED] will return playlist based in [data_added].
+  DATE_ADDED,
+}