+++ packages/on_audio_query/CHANGELOG.md
@@ -1,4 +1,22 @@
-## [[2.7.0](https://github.com/LucJosin/on_audio_query/releases/tag/2.7.0)] - [29.03.2023]
+## [[2.8.0](https://github.com/LucJosin/on_audio_query/releases/tag/2.8.0)]
+
+### Features
+
+- **Added** `showDetailedLog`.
+
+### Changes
+
+- **Moved** `android` and `ios` into separated folders.
+- **Replaced** `/details` with `/src`.
+
+### ⚠ Important Changes
+
+#### Android
+
+- **Updated** kotlin version from `1.4.32` to `1.6.10`. - [#110](https://github.com/LucJosin/on_audio_query/issues/110)
+- **Updated** kotlin coroutines version from `1.5.2-native-mt` to `1.6.4`.
+
+## [[2.7.0](https://github.com/LucJosin/on_audio_query/releases/tag/2.7.0)] - [03.29.2023]
 
 ### Features